chore: lib/types디렉토리 삭제, 루트에서 types디렉토리 관리하기로 통일 / mainProps관련 데이터 MainRomm컴포넌트로 이동
