--- lib/types/uploadFormProps.ts
@@ -1,5 +1,0 @@
-export interface UploadFormProps {
-  backgroundImage: string
-  videoId?: string
-  text?: string
-}
