--- lib/types/uploadPageProps.ts
@@ -1,4 +1,0 @@
-export type Props = {
-  url: string
-  hasImage: boolean
-}
